+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### 0.3.9
 * Added a validation that incident field can not be required.
+* Fixed an issue where *create-content-artifacts* did not copy TestPlaybooks in Packs without prefix of *playbook-*.
 * Added a validation the a playbook can not have a rolename set.
 * Added to the image validator the new DBot default image.
 * Added the fields: elasticcommonfields, quiet, quietmode to the Playbook schema.